Correctly initialize AF register
The union combining the A and F registers is broken which causes AF to not be equal to ((A <<8) | F). Instead, when addressing AF, A is an arbitrary number.
This is usually no big deal since the emulator always accesses either A or F individually. However, when the CPU gets reset, it initializes the combined AF register to 0x01B0 but because of this issue, A does not get initialized to 0x01 as it should. This causes A to not hold its correct startup value.
Games that depend on the AF register to hold 0x01B0 on CPU reset will then not work properly (e.g. Super Mario Land 2).